% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_est.R
\name{graph_est}
\alias{graph_est}
\title{Estimate graph from observational data}
\usage{
graph_est(X, method = "TD", ...)
}
\arguments{
\item{X}{a matrix or data frame containing the observed variables.}

\item{method}{}
}
\value{
WHAT IS RETURNED
}
\description{
SHORT DESCRIPTION
}
\details{
LONG DESCRIPTION
}
