% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\docType{data}
\name{LowDense}
\alias{LowDense}
\title{Data - Low dimensional dense simulation}
\format{a data frame with 8 variables:
\describe{
  \item{method}{the topological ordering method}
  \item{n}{number of observations}
  \item{p}{number of parameters}
  \item{sigma}{stnadard deviation of the common error}
  \item{Kendall}{Kendalls tau btween the true and esimated topological ordering}
  \item{Recall}{percent of arrows true positive arrows in estimated graph}
  \item{Flipped}{percent of arrows flipped compared to true graph}
  \item{FDR}{percent of arrows false positivesin the estimateed graph}
}}
\usage{
LowDense
}
\description{
A dataset containing the result of a large scale simulation of the low 
dimensional and dense setting.
}
\keyword{datasets}
