% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\docType{data}
\name{LowSim}
\alias{LowSim}
\title{Data - Low dimensional simulation}
\format{a data frame with 8 variables:
\describe{
  \item{p}{number of parameters}
  \item{graph_setting}{Low dimensional sparse, dense or high dimensional A/B}
  \item{l}{lower bound for absolute ausal effect}
  \item{u}{upper bound for sbaolute causal effect}
  \item{uniquq_ordering}{TRUE / FALSE}
  \item{n}{number of observations}
  \item{sigma}{standard deviation of the common error}
  \item{measure}{Model selection parameter, "mse", "mae", "deviance", "AIC", "BIC"}
  \item{which}{Model selection parameter, "min" or "1se"}
  \item{method}{"TD", "BU", "HTD", "HBU"}
  \item{max.degree}{HTD parameter}
  \item{search}{"full", "B&B" or "OMP"}
  \item{M}{Tuning parameter in organic lasso}
  \item{Kendall}{Kendalls tau btween the true and esimated topological ordering}
  \item{Hamming}{Structual Hamming Distance between true and estimated graph}
  \item{Recall}{percent of arrows true positive arrows in estimated graph}
  \item{Flipped}{percent of arrows flipped compared to true graph}
  \item{FDR}{percent of arrows false positivesin the estimateed graph}
}}
\usage{
LowSim
}
\description{
A dataset containing the result of a large scale simulation of the low 
dimensional setting.
}
\keyword{datasets}
