% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_graph_est.R
\name{sim_B}
\alias{sim_B}
\title{Simulates a graph}
\usage{
sim_B(p, graph_setting, l, u, unique_ordering = TRUE)
}
\arguments{
\item{p}{number of variables to simulate}

\item{graph_setting}{either "dense", "sparse", "A", or "B". These settings 
correspons to the settings described in the article.}

\item{l}{lower bound for direct abolute causal effect}

\item{u}{upper bound for direct absolute causal effect}

\item{unique_ordering}{if \code{TRUE} the simulated graph will have a unique 
topological ordering}
}
\value{
Returns a matrix of size \code{p} times \code{p} encoding a graph 
with regression coefficients.
}
\description{
This is an internal simulation function. It creates a graph with regression 
coefficients.
}
\details{
The \code{sim_B} function simulates a causal graph with regression 
coefficients. The graph is stored as matrix, where the (i,j)th entry in the
matrix is non zero only if there is an arrow from j to i in the graph. The 
value in the (i,j)th entry corresponds to the direct causal effect from j
to i.
}
\examples{
sim_B(5, "dense", 0.3, 1)

sim_B(5, "sparse", 0.7, 1, uniqe_ordering = FALSE)

}
