% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_graph_est.R
\name{sim_graph_est}
\alias{sim_graph_est}
\alias{sim_B}
\alias{sim_X}
\alias{standard_senarios}
\alias{plot_sim_analysis}
\alias{make_table}
\title{[A] Simulation tool for graph_est}
\usage{
sim_graph_est(scenarios, m)

sim_B(p, graph_setting, l, u)

sim_X(B, n, sigma, alpha = rep(1, ncol(B)))

standard_senarios(setting)

plot_sim_analysis(data)

make_table(data)
}
\arguments{
\item{m}{the number of simulations of each senario}

\item{senarios}{a data frame or matrix}
}
\value{
WHAT IT RETURNS
}
\description{
This is an internal simulation function for testing preformance of the 
\code{\link{graph_est}} function.
}
\details{
LONG DESCRIPTION
}
\examples{

}
