% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_graph_est.R
\name{sim_X}
\alias{sim_X}
\title{Simulates data from a graph}
\usage{
sim_X(B, n, sigma, alpha = rep(1, ncol(B)))
}
\arguments{
\item{B}{a upper triangular matrix encoding a graph}

\item{n}{number of observations from the graph}

\item{sigma}{the common standard error}

\item{alpha}{a vector of length equal to the number of parameters given by 
\code{B}. These \code{alpha}s are multiplied to the common \code{sigma}.}
}
\value{
The \code{sim_X} function returns a matrix with data from the graph 
\code{B} with columns equal to the number of columns in \code{B} and rows 
equal to \code{n}.
}
\description{
This is an internal simulation function for simulating data from a graph.
}
\details{
The \code{sim_B} function simulates a causal graph with regression 
coefficients. The graph must be encoded as an upper triangular matrix.
}
\examples{
(B <- matrix(c(0,0,0,1,0,0,1,1,0), ncol = 3))

sim_X(B, 10, 1)

sim_X(B, 10, 1, c(1, 5, 10))

}
