% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_order.R
\name{top_order}
\alias{top_order}
\title{Find the Topological Ordering of parameters}
\usage{
top_order(X, method = "TD", ...)
}
\arguments{
\item{X}{a matrix containing the oberved variables.}

\item{max.degree}{an integer larger then 1 describing the maximum in-degree 
in the underlying graph.}
}
\value{
a vector of length equal to the number of parameters indicating the 
  estimated topological ordering
}
\description{
Find the Topological Ordering of parameters, which are assumed to come from 
a linear SEM (also known as SCM) with equal error variance.
}
\details{
MORE DETAILED DECSRIPTION
}
\examples{
n <- 1000
p <- 5
sigma <- 1

# an example where all nodes in the graph are root nodes
X <- matrix(rnorm(n * p, 0, sigma), ncol = p)
order <- top_order(X)

# an example where pa(X_i) = X_j for all j < i
for (i in seq_len(p)[-1]) {
  X[ ,i] <- X[ ,i-1] + X[ ,i]
}
order <- top_order(X)


}
