% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_est.R
\name{graph_est}
\alias{graph_est}
\title{Estimate graph from observed data}
\usage{
graph_est(X, method = "TD", measure = "deviance", which = "1se", ...)
}
\arguments{
\item{X}{a matrix or data frame containing the observed variables.}

\item{method}{the estimation method. Posible choises are TD, BU, HTD}

\item{...}{terms passed to the method specific estimations steps.}
}
\value{
WHAT IS RETURNED
}
\description{
Estimates the graph of a linear SEM with assumed equal variance of the noise 
terms.
}
\details{
LONG DESCRIPTION
}
