% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_from_top.R
\name{graph_from_top}
\alias{graph_from_top}
\title{Find graph from topological ordering}
\usage{
graph_from_top(X, top)
}
\arguments{
\item{X}{a matrix containing the observed variables}

\item{top}{a topological ordering of the variables}
}
\value{
the B graph estimated from the data
}
\description{
SHORT DESCRIPTION
}
\details{
LONG DESCRIPTION
}
\examples{

# we create some data from the graph B
n <- 1000
B <- matrix(c(0,1,0,1,
              0,0,2,0,
              0,0,0,1,
              0,0,0,0), ncol = 4, nrow = 4, byrow = T)
X <- matrix(0, ncol = 4, nrow = n)
for (i in 1:4) {
  X[ ,i] <- X \%*\% B[ ,i] + rnorm(n)
}

# we then find the graph using a topological ordering
top <- c(1,2,3,4) # from B we know this to be the true ordering
grap_from_top(X, top)

}
