% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_order.R
\name{top_order}
\alias{top_order}
\title{Find the Topological Ordering of parameters}
\usage{
top_order(X, method = "TD", ...)
}
\arguments{
\item{X}{A matrix containing the oberved variables.}

\item{method}{The estimation method. Posible choises are "TD", "BU", "HTD"
and "HBU".}

\item{...}{Terms passed to the method specific estimation steps.

  If the method "HBU" is specified it is possible to specify a tuning 
  parameter \code{M} to be used in the organic lasso.

  If the method "HTD" is specified it is posible to specify a 
  \code{max.degree} and a \code{search}. The \code{max.degree} specifies the 
  assumed maximal in-degree in the true causal graph. The parameter 
  \code{search} may be set to either "full", "B&B" or "OMP" indicating how 
  the algorithem should search for \code{max.degree} number of parameters to 
  produce the lowest conditional variance.
  
  The "full" search method simply looks at all possible subsets of the 
  current ansestreal set of size \code{max.degree}, the "B&B" method searches
  via a bound and branch method, and lastly "OMP" uses orthogonal matching 
  pursuit (fprward selection) to find \code{max.degree} variables from the 
  ansestreal set.}
}
\value{
A vector of length equal to the number of parameters indicating the 
  estimated topological ordering is returned.
}
\description{
Find the Topological Ordering of parameters, which are assumed to come from 
a linear SEM (also known as SCM) where it is assumed that the independent 
noise terms have equal error variance.
}
\details{
MORE DETAILED DECSRIPTION
}
\examples{
n <- 1000
p <- 5
sigma <- 1

# an example where all nodes in the graph are root nodes
X <- matrix(rnorm(n * p, 0, sigma), ncol = p)
order <- top_order(X)

# an example where pa(X_i) = X_j for all j < i
for (i in seq_len(p)[-1]) {
  X[ ,i] <- X[ ,i-1] + X[ ,i]
}
order <- top_order(X, method = "TD")
order <- top_order(X, method = "BU")
order <- top_order(X, method = "HTD", max.degree = 2L, search = "B&B")
order <- top_order(X, method = "HBU")


}
\seealso{
The wrapper function \code{\link{graph_est}}combines \code{top_order} 
  and the model selection function \code{\link{graph_from_top}} into one 
  function that estimates the causal graph from the data.
}
